refactor: make global.error package
