+++ src/main/java/io/github/teamwaff1e/waffle/global/error/controlleradvice/ErrorControllerAdvice.java
@@ -1,9 +1,9 @@
-package io.github.teamwaff1e.waffle.global.exception.controlleradvice;
+package io.github.teamwaff1e.waffle.global.error.controlleradvice;
 
 import io.github.teamwaff1e.waffle.global.dto.response.ExceptionResponseDto;
+import io.github.teamwaff1e.waffle.global.error.code.ErrorCode;
+import io.github.teamwaff1e.waffle.global.error.controller.ErrorController;
 import io.github.teamwaff1e.waffle.global.exception.auth.WAuthenticationException;
-import io.github.teamwaff1e.waffle.global.exception.code.ErrorCode;
-import io.github.teamwaff1e.waffle.global.exception.controller.ErrorController;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
